style: move menu functions to menu.js
feat: add save and get localStorage of station
